feat: add Telegraph package and create migration files for telegraph_bots and telegraph_chats tables

+++ database/migrations/2025_06_12_054915_create_drivers_table.php
@@ -11,19 +11,21 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('drivers', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->string('full_name');
-            $table->string('status')->default('pending'); // pending, active, blocked
-            $table->string('license_number')->nullable();
-            $table->string('license_photo')->nullable();
-            $table->string('car_model')->nullable();
-            $table->string('car_photo')->nullable();
-            $table->string('country')->nullable();
-            $table->string('city')->nullable();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('drivers')) {
+            Schema::create('drivers', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+                $table->string('full_name');
+                $table->string('status')->default('pending'); // pending, active, blocked
+                $table->string('license_number')->nullable();
+                $table->string('license_photo')->nullable();
+                $table->string('car_model')->nullable();
+                $table->string('car_photo')->nullable();
+                $table->string('country')->nullable();
+                $table->string('city')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
