+++ database/migrations/2025_06_12_055014_create_clients_table.php
@@ -11,14 +11,17 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('clients', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->string('full_name');
-            $table->string('phone');
-            $table->string('telegram_id')->unique();
-            $table->timestamps();
-        });
+     
+        if (!Schema::hasTable('clients')) {
+            Schema::create('clients', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+                $table->string('full_name');
+                $table->string('phone');
+                $table->string('telegram_id')->unique();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
